+++ packages/gdl-frontend/lib/offlineLibrary/OfflineLibrary.js
@@ -2,10 +2,11 @@
 import { coverImageUrl } from 'gdl-image';
 import localForage from 'localforage';
 
-import type { BookDetails, Chapter } from '../types';
-import { fetchChapter } from '../fetch';
+import type { BookDetails, Chapter } from '../../types';
+import TimestampModel from './TimestampModel';
+import { keyForBook, CACHE_NAME } from './index';
+import { fetchChapter } from '../../fetch';
 
-const CACHE_NAME = 'gdl-offline';
 // 7 days
 const MAX_AGE_MS = 7 * 24 * 60 * 60 * 1000;
 
@@ -13,42 +14,7 @@ function openCache() {
   return window.caches.open(CACHE_NAME);
 }
 
-/**
- * We keep track of when the book was offlined, so we can expire it after a certain time has passed
- */
-export class TimestampModel {
-  timestampStore = localForage.createInstance({
-    name: CACHE_NAME,
-    storeName: 'timestamp'
-  });
-
-  getTimestamp = async (id: number | string, language: string) =>
-    this.timestampStore.getItem(keyForBook(id, language));
-
-  setTimestamp = async (id: number | string, language: string) =>
-    this.timestampStore.setItem(keyForBook(id, language), Date.now());
-
-  deleteTimestamp = async (id: number | string, language: string) =>
-    this.timestampStore.removeItem(keyForBook(id, language));
-
-  getTimeStamps = async () => {
-    const timestamps = [];
-
-    await this.timestampStore.iterate(
-      (timestamp: number, compositeId: string) => {
-        // Extract the id and language code from the composite id
-        const [id, language] = compositeId.split(/-(.+)/);
-        timestamps.push({ id, timestamp, language });
-      }
-    );
-
-    return timestamps;
-  };
-
-  clear = () => this.timestampStore.clear();
-}
-
-export class OfflineLibrary {
+export default class OfflineLibrary {
   bookStore = localForage.createInstance({
     name: CACHE_NAME,
     storeName: 'books'
@@ -215,32 +181,6 @@ export class OfflineLibrary {
 }
 
 /**
- * Singleton that is null unless the client has offline support
- */
-const offlineLibrary =
-  typeof window !== 'undefined' &&
-  'serviceWorker' in navigator &&
-  typeof navigator.onLine === 'boolean'
-    ? new OfflineLibrary()
-    : null;
-
-export default offlineLibrary;
-
-/**
- * Book ids aren't unique. So we make a composite key together with the language
- */
-function keyForBook(
-  bookOrId: string | number | BookDetails,
-  language?: string
-) {
-  return arguments.length > 1
-    ? // $FlowFixMe
-      `${bookOrId}-${language}`
-    : // $FlowFixMe
-      `${bookOrId.id}-${bookOrId.language.code}`;
-}
-
-/**
  * Get all unique image URLs in a book and it's chapters
  */
 function getImageUrls(book: BookDetails, chapters: Array<Chapter>) {
@@ -256,9 +196,3 @@ function getImageUrls(book: BookDetails, chapters: Array<Chapter>) {
   }
   return imageUrls;
 }
-
-/**
- * Make sure we're running a service worker in the top scope
- */
-export const runsServiceWorker = async () =>
-  Boolean(await window.navigator.serviceWorker.getRegistration('/'));
