+++ packages/gdl-frontend/lib/offlineLibrary/index.js
@@ -1,0 +1,32 @@
+// @flow
+import type { BookDetails } from '../../types';
+import OfflineLibrary from './OfflineLibrary';
+
+export const CACHE_NAME = 'gdl-offline';
+
+/**
+ * Book ids aren't unique. So we make a composite key together with the language
+ */
+export function keyForBook(
+  bookOrId: string | number | BookDetails,
+  language?: string
+) {
+  return arguments.length > 1
+    ? // $FlowFixMe
+      `${bookOrId}-${language}`
+    : // $FlowFixMe
+      `${bookOrId.id}-${bookOrId.language.code}`;
+}
+
+const serviceworker =
+  typeof window !== 'undefined' && 'serviceWorker' in navigator;
+
+const online =
+  typeof window !== 'undefined' && typeof navigator.onLine === 'boolean';
+
+/**
+ * Singleton that is null unless the client has offline support
+ */
+
+const offlineLibrary = serviceworker && online ? new OfflineLibrary() : null;
+export default offlineLibrary;
